+++ src/Sentry/ITransaction.cs
@@ -14,6 +14,12 @@ namespace Sentry
         new string Name { get; set; }
 
         /// <summary>
+        /// Whether the parent transaction of this transaction has been sampled.
+        /// </summary>
+        // 'new' because it adds a setter
+        new bool? IsParentSampled { get; set; }
+
+        /// <summary>
         /// Flat list of spans within this transaction.
         /// </summary>
         IReadOnlyCollection<ISpan> Spans { get; }